+++ R/convert_odds_to_d.R
@@ -21,7 +21,9 @@ convert_odds_to_d <- odds_to_d
 
 #' @rdname d_to_r
 #' @export
-logodds_to_d <- function(odds, log = TRUE, ...) odds_to_d(odds, log = log, ...)
+logodds_to_d <- function(odds, log = TRUE, ...) {
+  odds_to_d(odds, log = log, ...)
+}
 
 
 
@@ -47,7 +49,7 @@ convert_d_to_odds <- d_to_odds
 #' @rdname d_to_r
 #' @export
 odds_to_r <- function(odds, log = FALSE, ...) {
-  d_to_r(convert_odds_to_d(odds, log = log))
+  d_to_r(odds_to_d(odds, log = log))
 }
 
 #' @rdname d_to_r
@@ -56,14 +58,16 @@ convert_odds_to_r <- odds_to_r
 
 #' @rdname d_to_r
 #' @export
-logodds_to_r <- function(odds, log = TRUE, ...) odds_to_r(odds, log = log, ...)
+logodds_to_r <- function(odds, log = TRUE, ...) {
+  odds_to_r(odds, log = log, ...)
+}
 
 
 
 #' @rdname d_to_r
 #' @export
 r_to_odds <- function(r, log = FALSE, ...) {
-  d_to_odds(convert_r_to_d(r), log = log)
+  d_to_odds(r_to_d(r), log = log)
 }
 
 #' @rdname d_to_r