+++ R/convert_tFz_to_r.R
@@ -105,6 +105,7 @@ t_to_r <- function(t, df_error, ci = 0.95, ...) {
 
 
 #' @rdname t_to_r
+#' @importFrom stats qnorm
 #' @export
 z_to_r <- function(z, n, ci = 0.95, ...) {
 
@@ -117,7 +118,7 @@ z_to_r <- function(z, n, ci = 0.95, ...) {
     alpha <- 1 - ci
     probs <- c(alpha / 2, 1 - alpha / 2)
 
-    qs <- qnorm(probs)
+    qs <- stats::qnorm(probs)
     zs <- cbind(qs[1] + z, qs[2] + z)
 
     res$CI_low <- zs[,1] / sqrt(zs[,1]^2 + n)