+++ R/print.effectsize_table.R
@@ -23,6 +23,11 @@ print.effectsize_table <- function(x, digits = 2, ...){
 
   cat(insight::format_table(x, digits = digits))
 
+  if (!is.null(method <- attr(x_orig, "std_method"))) {
+    method <- paste0(toupper(substr(method, 1L, 1L)), substr(method, 2L, nchar(method)))
+    insight::print_color(color = "blue", paste0("\n# Standardization method: ", method, "\n"))
+  }
+
   invisible(x_orig)
 }