+++ tests/testthat/test-convert.R
@@ -1,22 +1,27 @@
 if (require("testthat") && require("effectsize")) {
   test_that("odds_to_probs", {
-    testthat::expect_equal(odds_to_probs(-1.6), 2.66, tolerance = 0.01)
-    testthat::expect_equal(odds_to_probs(-1.6, log = TRUE), 0.17, tolerance = 0.01)
-    testthat::expect_equal(probs_to_odds(2.66), -1.6, tolerance = 0.01)
-    testthat::expect_equal(probs_to_odds(0.17, log = TRUE), -1.6, tolerance = 0.01)
-    testthat::expect_equal(odds_to_probs(-1.6, log = TRUE), 0.17, tolerance = 0.01)
+    testthat::expect_equal(odds_to_probs(3), 0.75, tolerance = 0.01)
+    testthat::expect_equal(probs_to_odds(0.75), 3, tolerance = 0.01)
+    testthat::expect_equal(probs_to_odds(0.75, log = TRUE), 1.098, tolerance = 0.01)
+    testthat::expect_equal(odds_to_probs(1.098, log = TRUE), 0.75, tolerance = 0.01)
 
-    testthat::expect_true(
-      ncol(odds_to_probs(
+
+
+    testthat::expect_equal(
+      ncol(df <- odds_to_probs(
         iris,
         select = c("Sepal.Length"),
-        exclude = c("Petal.Length")
-      )) == ncol(probs_to_odds(
-        iris,
+        exclude = c("Petal.Length"),
+        log = TRUE
+      )), 5)
+
+    testthat::expect_equal(
+      ncol(probs_to_odds(
+        df,
         select = c("Sepal.Length"),
-        exclude = c("Petal.Length")
-      ))
-    )
+        exclude = c("Petal.Length"),
+        log = TRUE
+      )), 5)
   })
 
   test_that("odds_to_d", {
@@ -33,15 +38,5 @@ if (require("testthat") && require("effectsize")) {
 
     testthat::expect_equal(odds_to_r(odds = d_to_odds(d = r_to_d(0.5))), 0.5, tol = 0.001)
     testthat::expect_equal(odds_to_d(r_to_odds(d_to_r(d = 1), log = TRUE), log = TRUE), 1, tolerance = 0.001)
-  })
-
-  test_that("d_to_percentage", {
-    testthat::expect_equal(d_to_percentage(d = 1), 0.162, tolerance = 0.01)
-    testthat::expect_equal(percentage_to_d(percentage = 0.01), 0.0618, tolerance = 0.01)
-  })
-
-
-  test_that("z_to_percentile", {
-    testthat::expect_equal(z_to_percentile(percentile_to_z(0.975)), 0.975, tolerance = 0.001)
   })
 }